+++ cmd/datawatch/main.go
@@ -53,7 +53,7 @@ import (
 )
 
 // Version is set at build time via -ldflags.
-var Version = "0.5.0"
+var Version = "0.5.1"
 
 var (
 	cfgPath    string